feat(factors): add support for prime factorization
This commit adds support for prime factorization to the factors program. This means that the program can now factor numbers into a product of two prime numbers.

The new functionality is implemented in the `factor()` function, which takes a number as input and returns a tuple of the two prime factors. The function uses the Sieve of Eratosthenes algorithm to find all of the prime numbers up to the square root of the input number. It then divides the input number by each of the prime numbers in turn, until the number is reduced to 1. The two prime factors are the last two numbers that the input number is divided by.

The `factor()` function is now used by the `main()` function to factor all of the numbers in the input file. The results of the factorization are printed to the standard output.

This commit also includes a number of unit tests to verify the functionality of the `factor()` function.